Update patch set 1
Patch Set 1:

> You've made the same change twice, exactly same code duplicated. 
 > Can you factor it out so that the choice of name is made in exactly
 > one place?

Both the file's are coming from two different project's. So no need to refactor it.

 > You've made the same change twice, exactly same code duplicated. 
 > Can you factor it out so that the choice of name is made in exactly
 > one place?


Both the file's are belonging to two separate project's. So no need to refactor it on a same place

 > You've made the same change twice, exactly same code duplicated. 
 > Can you factor it out so that the choice of name is made in exactly
 > one place?


Both the file's belonging to two separate project's. So no need to refactor it on the same place.

Patch-set: 1